stack/0503_next_greater_element_ii/solution_test.go
package main

import (
	"reflect"
	"testing"
)

func Test_nextGreaterElements(t *testing.T) {
	type args struct {
		nums []int
	}
	tests := []struct {
		name string
		args args
		want []int
	}{
		{
			"example1",
			args{
				nums: []int{1, 2, 1},
			},
			[]int{2, -1, 2},
		},
		{
			"example2",
			args{
				nums: []int{1, 2, 3, 4, 3},
			},
			[]int{2, 3, 4, -1, 4},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := nextGreaterElements(tt.args.nums); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("nextGreaterElements() = %v, want %v", got, tt.want)
			}
		})
	}
}
